initial setup
This commit includes functionality to write an initial config file when
running on a storage location. It will prompt for an initial feed and will
set up a twitter application and access token to tweet as a particular user.

Tweeting was pulled off of the Diff class and turned into a function.
Since the twitter client needs to be configured with access tokens appropriate
to a feed it is built up anew each time. The reason for this is that
diffmachine might be monitoring more than one feed and want to tweet on accounts
that are specific to each feed. For example if diffmachine was watching a feed
for the Washington Post and the LA Times.

The EntryVersion object now has diff and next_diff methods that hopefully make
it easier to navigate to diffs relevant for a given version.