normalize text
Extracted the text cleaning code to a separate function `_normalize` which is
used when comparing what was just received (after readability and bleach) to
what was stored in the db.

It is important to reapply _normal to what is stored in the database in case the
rules around how to normalize text are modified in the future.

Fixes #14