lookup latest version with URL
This is a small but important change to make sure that the latest version
is looked up using the final URL for an entry. It appears that some newspapers
will have multiple feeds that contain different entry URLs that redirect to
the same URL.

For example:

    Feed1 -> Entry1 -> Page1
    Feed2 -> Entry2 -> Page1

So it is imporant when looking up the latest version for an entry to use the
canonical URL, or the URL after all the redirects are done. Otherwise the page
will be reported multiple times as having changed, when it really only changed
once.

Fixes #17