doublecheck htmldiff
This commit includes an extra check in Entry.get_latest to make sure that the
generated diff has either an <ins> or <del> in it. If it doesn't that means
while the readability summaries are different the resulting diff didn't catch
any changes. This can happen with <br> and other whitespace changes.

This change also explicitly sets the coding on diffengine.__init__. Fixes #24